Change an int to a ptrdiff_t

+++ modem.c
@@ -79,7 +79,7 @@ static bool menu_find_url(url_t url_buffer, const page_t menu_page,
   const char *url_end = strstr(url_begin, URL_ENDING);
   if (!url_end) return false;
 
-  const int url_length = url_end - url_begin;
+  const ptrdiff_t url_length = url_end - url_begin;
 
   // Add 1 for null termination.
   string_copy(url_begin, url_buffer, min(url_length + 1, MAX_URL_LENGTH));
